pyproject.toml
[project]
name = "qissuereporter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "pillow>=11.3.0",
    "pydantic>=2.11.9",
    "pyqt6>=6.9.1",
    "qcustomwidgets",
    "qcustomwindow",
]

[tool.uv.sources]
qcustomwindow = { git = "https://github.com/CrinitusFeles/QCustomWindow" }
qcustomwidgets = { git = "https://github.com/CrinitusFeles/QCustomWidgets" }
